Rem U04 + Edit /125 + Add U04 stat on Ebay crispychris + Add date on L01
Unknown status U04 is sticker /125. The Ebay line after that line, crispychrisb, didn't have a unknown status before I added it in there now. Because of the fact that the original U04 status is now removed because it is solved and the line after didn't have that status as it should of have, I gave it the U04 status to replace it. Also added in date on Ebay L01 from dragonslayer875 :)