Adding Twitter U27 MinnMaxShow and U28 DaniaGames
See sticker list U27 and U28 for full details. Found via Twitter search. It is currently unknown who the current owner of U27 MinnMaxShow's edition is, considering it was won by someone that has yet to be identified. U28 DaniaGames received their edition from Playdead as a gift/donation, but by looking at the images, it is impossible to say yet if the sticker is disposed of or not.