Updates on U28 (L34), U11, U12, U14, U16, U27
Some recent updates on unknwon statuses: Unknown 28 confirmed lost. New messages sent to Unknown 11, Unknown 12, Unknown 14. Unknown 16 is resolved, but remains with the Unknown status, see description in Unknown 16 for explanation. For Unknown 27 I'm still awaiting reply from the winner, but updated it to the latest occurence of anything happening.